fix!: prefer _Sancta_ over _Sacra_.
This is the more usual phrase. I suspect "sacra" was a typo at the time. It's
not *wrong* (although it would more usually come after the noun), but there's
little reason to prefer it.

The meaning is more or less the same. IMHO (and I'm no latinist) _sacra_ has
more of an overtone of "the sacred things" (cf. Musica Sacra = church music)
where _sancta_ simply means "holy" (Sancta Maria). But even that is probably
purely contextual. There is a _Missa Sacra_ by Schumann, but I don't know it, so
that is unlikely to be the root of the confusion.